ASoC: soc-compress: Add support for not memory mapped DSPs
The ASoC compressed API did not implement the copy callback in its
compressed ops which is required for DSPs that are not memory mapped.

This patch creates a local copy of the compress ops for each runtime and
modifies them with a copy callback as appropriate.

Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>
Acked-by: Vinod Koul <vinod.koul@intel.com>
Signed-off-by: Mark Brown <broonie@opensource.wolfsonmicro.com>
(cherry picked from commit 1f88eb0f0660f8b58a1fe9011f3d3a350c7dd194)

Change-Id: I4d787f41ce25cba496004f46fe24fb075833eca0
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>