regmap: debugfs: Simplify calculation of `c->max_reg'
We don't need to use any of the file position information
to calculate the base and max register of each block.  Just
use the counter directly.

The `i = base' at the top to avoid GCC flow analysis bugs.  The
value of `i' can never be undefined or 0 in the if (c) { ... }.

Change-Id: Id6080b753952f2805c3642793b7db34e1a4c70e2
Signed-off-by: Dimitris Papastamos <dp@opensource.wolfsonmicro.com>