mfd: arizona: Try to use interrupt flags from interrupt controller
If no irq_flags are passed in platform data then query the interrupt
controller for the trigger type and try to use that. This provides
default operation with a wider range of hardware and will be needed
for device tree support where the interrupt flags are configured on
the interrupt controller.

Change-Id: I49be9b79543c331cde603172d4e53bb2e8cbe587
Signed-off-by: Mark Brown <broonie@opensource.wolfsonmicro.com>