ASoC: wm_adsp: Ensure we don't attempt to create controls twice
wm_adsp_create_grouped_control is called from a loop which iterates
through the alg_regions list. However it also adds new regions into this
list, controls for these regions will be added twice, once when the new
region is added to the list and once when the iteration reached the new
region in the list. This patch changes this such that the control is
only added once it is processed from the alg_regions list.

Change-Id: If8d30f5dc1ce1b496c198e31cf7e3b350c646d25
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>