ASoC: arizona: Calculate FLL gain last
No part of the FLL calculation depends on the value determined for the
gain but the gain does depend on other values. In preparation for future
updates this patch moves the gain to be the last thing calculated.

Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>
Signed-off-by: Mark Brown <broonie@linaro.org>
(cherry picked from commit 5a3935c7643966e4172e7a704a48a35f9b4dc668)

Change-Id: I1feaee217dee944627cffa37c956d34a1255da9d
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>