mfd: wm5102: Remove cache_bypass from manual register patching
On the wm5102 the register patches are applied manually, rather than by
the regmap core. This application is wrapped in calls to
regcache_cache_bypass. However, this is dangerous as other threads may
be accessing the hardware at the same time as the pm_runtime operations
and if they do so during the period whilst cache_bypass is enabled those
writes will miss the cache when they shouldn't.

As the cache_bypass is not strictly necessary for applying the patches
(the device is always powered on whilst it is appliedi, and all we loose
is adding a couple of registers to the cache) remove the calls to
regcache_cache_bypass to avoid this situation.

Change-Id: Ib58ac37c945a21ce03f308e6ffad346bbf11d1fe
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>