mfd: arizona: Use new regmap features for manual register patch
On the wm5102 the register patches are applied manually, rather than by
the regmap core. This application is wrapped in calls to
regcache_cache_bypass. However, this is dangerous as other threads may
be accessing the hardware at the same time as the pm_runtime operations
and if they do so during the period whilst cache_bypass is enabled those
writes will miss the cache when they shouldn't.

Apply the register patch using the new regmap_multi_reg_write_bypassed
function to avoid this problem. Also remove the call to
regcache_cache_bypass from the hardware patch application as it is
unneeded there and creates a similar window for writes to miss the
cache.

Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>
Acked-by: Lee Jones <lee.jones@linaro.org>
Signed-off-by: Mark Brown <broonie@linaro.org>
(cherry picked from commit 1c18d2ca104c36fc2ce147cce053c62f61d2ea68)

Change-Id: I848f2ef78d9c1a0875fff332eee9b002e08ebd52
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>