ASoC: wm_adsp: Change firmware preloader widget to be a PGA
The snd_soc_dapm_dai_link widget type does not exist in a stock 3.4
kernel so we shouldn't be using it for the preloader widget. Change this
back to being a PGA, it is still done as a pre-pmu event so should still
overlap with the output enables which is were most of the performance
gain comes from.

Change-Id: Ie7fd98081b2400ce9e64f5dde989f937d54809f6
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>