ASoC: florida: Keep track of if ez2control is forced on
If ez2control is forced on we shouldn't clear the triggered flag when we
close the stream as we may get another IRQ after the stream closes. This
patch adds a forced flag to keep track of when the DSP is forced on and
only clears trig when closing the stream if it is not forced. In the
case it is forced the virtual output power event will clear the trigger.

Change-Id: I8bd20db100c81bbc29411636d978aa7d75bfbb5c
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>
Signed-off-by: Nikesh Oswal <Nikesh.Oswal@wolfsonmicro.com>
Signed-off-by: Mayuresh Kulkarni <mkulkarni@opensource.wolfsonmicro.com>