ASoC: florida: Keep track of if ez2control is forced on
If ez2control is forced on we shouldn't clear the triggered flag when we
close the stream as we may get another IRQ after the stream closes. This
patch adds a forced flag to keep track of when the DSP is forced on and
only clears trig when closing the stream if it is not forced. In the
case it is forced the virtual output power event will clear the trigger.

Change-Id: I580ec419a892bfa23724503c64750127b4d2f434
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>