ASoC: arizona: Revert manual speaker enable
The OUT4L and OUT4R widgets are not registers for PRE_PMU or POST_PMD
events, as such the manual speaker enable on wm5102 does not run even
close to correctly. As this has gone unoticed for a long time, seems
that the issue this was put into fix is rather a moot point. This
patch removes the manual speaker enable sequence for wm5102.

Change-Id: Id6581d8ca4bbb6156adb3d9d31cf282d406f0b06
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>