ASoC: wm5102: Add controls to allow shaping of ultrasonic response
Add controls to allow custom shaping of the ultrasonic response. This
patch also allows this to be turned on/off at runtime rather than
through pdata as was previously the case. Although, it should be noted
that settings will not affect a currently open audio stream, they will
be applied when the next audio stream is started.

Change-Id: If423cbe6d16a7683d113e24a5ae37196feed4b40
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>