switch: arizona: Get MICVDD against switch device
Previously we would do a regulator get against the main Arizona device
to obtain the MICVDD regulator. Arizona is an MFD device and normally
MICVDD will be supplied by one of its children (the arizona-micsupp
regulator). As devres destruction for the MFD device will run after all
its children have been destroyed, the regulator will be destroyed before
devres calls regulator_put. This causes a warning from both the
destruction of the child node, as the regulator is still open, and from
the put of the regulator as the regulator device has already been
destroyed.

A simple fix here is to get the regulator against the switch device
itself such that devres runs when the child is destroyed. This has the
additional benefit that if for some reason the switch driver is unloaded
the regulator reference won't hang around until the MFD is unloaded.

Change-Id: Ic38189eea2133a1da4ba5809ce66cf89913242c2
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>
Signed-off-by: Mayuresh Kulkarni <mkulkarni@opensource.wolfsonmicro.com>