mfd: arizona: Remove bogus warning in arizona_dvfs_down
The flags are simple booleans per block, not reference counts.
Each block that has a DVFS dependency is not itself reference
counted, it's a simple either/or case which these flags match.
Because the requestors don't reference count themselves we
allow them to notify the same state that they are already in.

Change-Id: I3ea6dbf3f9e265cfed0d3ca4b27ea4f757c90d6a
Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>
Signed-off-by: Mayuresh Kulkarni <mkulkarni@opensource.wolfsonmicro.com>