ASoC: wm_adsp: Factor out firmware feature checks
Create a new set of firmware feature flags that are initialized
when the firmware is powered up and check these where appropriate
instead of directly checking the firmware id. The flags are
currently still initialized based on firmware id but this does
at least abstract the checking of the feature from how it is
determined.

Change-Id: I208b62c6f29de22976278e17e5377aebc6335e59
Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>