ASoC: wm8285: Initial driver for WM8285/WM1840
The WM8285 is a highly-integrated low-power audio and sensor hub system
for smartphones, tablets and other portable audio devices.

IN1L and IN2 are multiplexed on wm8285, this patch only adds basic
support for this. Support for having and analogue and a digital input
connected to the same muxed input is still missing. But this enables
the common use-cases of a digital input or two analogue inputs.

Change-Id: I2a0b00f1bb7d92aaea26191d5910790c6461274b
Signed-off-by: Nariman Poushin <nariman@opensource.wolfsonmicro.com>
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>
Signed-off-by: Mayuresh Kulkarni <mkulkarni@opensource.wolfsonmicro.com>