ASoC: arizona: fix compressed stream IRQ race
The codec IRQ handlers were using the existence of the compressed
stream to decide whether to call the ADSP buffer code. But this
isn't correct, since we also have to allocate the ADSP buffers.
This gave a race condition that could cause a NULL pointer deref
if an IRQ happened between opening the compressed stream and
calling _set_params() to allocate the buffer.

Change-Id: I62a9d0bf50c029214f0e6dfa26074edc925e66a4
Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>
Signed-off-by: Mayuresh Kulkarni <mkulkarni@opensource.wolfsonmicro.com>