switch: arizona: handle button press for antenna state machine
For antenna state machine there are 3 manual debounce values
for plugin <wlf,antenna-manual-debounce>, plugout
<wlf,antenna-manual-db-plugout> and button
<wlf,micd-manual-debounce>.

The plugout debounce value will be used as a common debounce
for both button presses and to handle slow plugouts in cases
where a 4 pole is plugged into an antenna. This is done because
to support slow plugouts we need a larger debounce as compared
to button presses. But in cases where a 4 pole is directly
plugged in we use the dedicated button debounce.

Change-Id: I12251798b86233a237a5b3cf7af09259f19ada73
Signed-off-by: Nikesh Oswal <Nikesh.Oswal@wolfsonmicro.com>