mfd: clearwater: Correct application of 32-bit register patch
The 32-bit register patch is needlessly applied bypassing the normal
register patch functionality. This patch updates this to actually use
the register patch functionality. In the process this fixes a bug where
the 32-bit register patch function was being called before the chip
revision had been read causing the rev A patch to be applied to all
versions of the chip.

Change-Id: Ia304cc69a238846bc62e31265494a32b13e3a285
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>