regmap: Fix accidental unmasking of IRQs
The startup code in regmap_add_irq_chip() was trying to
initialize all interrupts to masked, but had the side-effect
of unmasking all interrupts that aren't defined in the
irq table.

Change this to use regmap_update_bits so that only the mask bits
for interrupts defined in the table are changed.

Change-Id: I95eb85ebafe3d542d2638d7f0d9dffd8c6e00494
Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>