ASoC: arizona: tie compr_info to compressed DAI
To allow for multiple compressed streams, have one instance
of the compr_info struct for each available compressed DAI
instead of sharing one. Each compr_info is tied to a specific
DAI and ADSP because we need to be able to service interrupts
from the firmware even if the compressed stream has not been
opened. There is no advantage to dynamically allocating
compr_info to compressed streams because we need to have exactly
the number of compr_info as compressed DAIs to be able to support
running all compressed DAIs at the same time so they may as well
be tied to each DAI.

Change-Id: Ie27ea4292716f9dd83aefe9fe4c57a229e2ce190
Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>