switch: arizona: Validate button press
when a 4 pole is slowly plugged out from antenna cable the hp
pins of 4 pole can touch the mic pins of antenna and we may
report a false button press this patch prevents it by
differentiating between hp and button impedance.

Differentiation is possible since the patch adds capability to
measure an impedance band around the hp impedance level and when
the state machine in button detection state hits this band it
considers it as a false button press which might have triggered
due to a slow plugout.

One tradeoff here is the button detection around this impedance
band will be disabled. So if a button (in a specific headset)
falls in the disabled impedance range then such a button won't
work (the headset as such and its other buttons would still work).
But the impedance range here is kept as small as possible by
the patch so that reduces the probability of hitting such a
condition. Speicfically for samsung headsets the impedance
of all the buttons fall outside this range so all functionality
works fine for samsung headsets.

Change-Id: Id5c2bd880aab5cbf3c20757563013d33475f59d2
Signed-off-by: Nikesh Oswal <Nikesh.Oswal@wolfsonmicro.com>