regmap: regcache-rbtree: Fix present bitmap resize
When inserting a new register into a block at the lower end the present
bitmap is currently shifted into the wrong direction. The effect of this is
that the bitmap becomes corrupted and registers which are present might be
reported as not present and vice versa.

Fix this by shifting left rather than right.

Fixes: 472fdec7380c("regmap: rbtree: Reduce number of nodes, take 2")
Reported-by: Daniel Baluta <daniel.baluta@gmail.com>
Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>
Signed-off-by: Mark Brown <broonie@kernel.org>
Cc: stable@vger.kernel.org
(cherry picked from commit 328f494d95aac8bd4896aea2328bc281053bcb71)

Change-Id: I3bc8b32546371bd62c5eb972a1233e8915534ad9
Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>