ASoC: wm_adsp: Remove private field from wm_coeff_ctl
The private field in wm_coeff_ctl is currently unused and given the
controls are entirely handled within the ADSP code it is not clear what
it would be used for in the future. Remove the field for now it can be
readded if it is ever required.

Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit ec184cfcb9303dd2e8620a2db902dd64e477f229)

Change-Id: I8a76ce537283e2327a088fc58105a6b768e4797b
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>