ASoC: wm_adsp: Improve variable naming
We have wm_adsp_region, wm_adsp_alg_region, and wmfw_region, the
variables for which are all frequently called region, this can get quite
confusing when review the code especially given some functions are quite
long. Consistently use mem for wm_adsp_regions, alg_region for
wm_adsp_alg_region and region for wmfw_region.

Whilst we are at it, we use a mix of adsp and dsp for pointers to the
wm_adsp structure standardise this on dsp.

Change-Id: Ib7b786742501abe64eefb29a92e832a5d28b3188
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>
Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>