switch: arizona: handle hp impedance measurement for moon
For moon we need to specify the sense and ground pins for headphone
impedance measurement. We also need to specify the output which will
be clamped when measuring the impedance. This is unlike previous
arizona codecs where we just specified the acc_det_mode as HPDETL
or HPDETR and clamping/unclamping was also automatic when impedance
measurement is activated.

Change-Id: Ibe406efbc5aec4fa03da5b2eec2a54f43a8cfc60
Signed-off-by: Nikesh Oswal <Nikesh.Oswal@wolfsonmicro.com>
Signed-off-by: Ammar Zahid Ali Syed <aali@opensource.wolfsonmicro.com>