switch: arizona: handle hp impedance measurement for moon
For moon we need to specify the sense and ground pins for headphone
impedance measurement. We also need to specify the output which will
be clamped when measuring the impedance. This is unlike previous
arizona codecs where we just specified the acc_det_mode as HPDETL
or HPDETR and clamping/unclamping was also automatic when impedance
measurement is activated.

Change-Id: I4465f0e1a23a1a6a4a2c1256c03ce1372eae317d
Signed-off-by: Nikesh Oswal <Nikesh.Oswal@wolfsonmicro.com>