mfd: arizona: Reorder initialisation of pm_runtime
The regulator enable in arizona_dev_init needs to be balanced with a
disable. This is currently done at the end of arizona_dev_init, with a
call to regulator_disable. However, if nothing has claimed a runtime
reference by that point this will actually disable DCVDD putting the
CODEC into suspend. However, the additional operations required to put
the CODEC into suspend are not honoured in this case.

As the PM runtime expects the device to be in a suspended state when it
takes charge, fix this issue by calling arizona_runtime_suspend before
we enable the PM runtime which will disable the regulator in a safe
manor.

Change-Id: Icee89a3fda5fa55bacd2f2dc57d42dee795fb3bc
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>