mfd: wm5102: Ensure we always boot the device fully
The wm5102 uses a custom boot sequence and the standard boot sequence is
disabled. However, the standard boot sequence must be run at least once
after a cold boot (caused by either a hard reset or removal of AVDD).
Unfortunately the register WRITE_SEQUENCER_CTRL_3 is not affected by the
hardware reset. This means if the device has been previously booted but
the AVDD supply has never been removed, arizona_dev_init will reset the
chip but the boot sequence will not run, which can cause numerous
problems. The solution is to manually clear this register and then boot
the chip again, which is what this patch does.

Change-Id: Ia38918ec2b840de53a6d08cfffdc18c68eabbca7
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>