ASoC: florida: Coalesce volume ramp sequence delays
Before powering down the headphones whilst DRE is not enabled we should
ramp down the volume for the output to ensure glitch free operation.
Currently, two 10mS delays are used for this. This patch coalesces these
with the standard power down delay. This is acceptable because the write
sequencer in the chip will queue all the write sequences, so the
headphone disable sequence can't run until the matching volume ramp has
completed. So in the driver we only need to allow enough total delay to
ensure that all the write sequences have finished.

Change-Id: I2e49891921ed4b51278d491629f155ce8e215644
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>