ASoC: wm_adsp: Move DVFS control into codec driver
In theory the ADSP driver should not need to know anything
about the codec it is part of. But the WM5102 needs DVFS
control based on ADSP clocking speed. This was being handled
by bundling part of the knowledge of this into the ADSP driver.

This change moves this handling out of the ADSP driver and
into the WM5102 driver.

Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 81ac58b13f815d7c7838bc347dd5d102707a11b7)

Change-Id: Iefdb9ffa8927c365aeb794d4117f1761f33eef11
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>