ASoC: florida: Remove PGA updates for headphone output disable
A DAPM sequence can't abort half way through, nothing else will be
attempting to alter the PGA volume during the sequence. This means that
at the end of the DAPM sequence both the cache and chip will have a
value of 0x80 for the PGA volume and there is no need to keep the cache
in sync during the sequence. This patch removes the additional writes
that were keeping the cache in sync.

Change-Id: Ia9668f9076ac3665382b3eaa39ee80033d0577b5
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>