mfd: arizona: Fix initialisation of the PM runtime
The PM runtime core by default assumes a chip is suspended when runtime
PM is enabled. Currently the arizona driver probe forcibly calls
arizona_runtime_suspend and then enables runtime PM. This is an ugly
workaround.

This patch adds a call to pm_runtime_set_active to inform the PM core
that the device is starting up active and moves the PM enabling to
around the IRQ initialisation to avoid any PM callbacks happening until
the IRQs are initialised so that it is safe for PM operation to
enable/disable the IRQ.

Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>
Signed-off-by: Lee Jones <lee.jones@linaro.org>
(cherry picked from commit 72e43164fd472f6c2659c8313b87da962322dbcf)

Change-Id: I21bef7499f996d0d935012d6f1cabf836388a49d
Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>