ASoC: arizona: Remove pointless readback in clearwater_spin_sysclk
The toggling and reading back of a value is only done under the
misguided notion that the SPI subsystem will allow writes/reads to be
reordered and that the writes may not have completed by the time the
function returns. As these assumptions are blatently false simplify
spin_sysclk to just read the chip ID register.

Change-Id: I7131994992a657559092928cf7a233b466e496e6
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>
Signed-off-by: Ammar Zahid Ali Syed <aali@opensource.wolfsonmicro.com>