ASoC: arizona: Remove pointless readback in clearwater_spin_sysclk
The toggling and reading back of a value is only done under the
misguided notion that the SPI subsystem will allow writes/reads to be
reordered and that the writes may not have completed by the time the
function returns. As these assumptions are blatently false simplify
spin_sysclk to just read the chip ID register.

Change-Id: Ib3bae9b51d46981e09f4d2c3e6e1a8bd97b94ec4
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>
Signed-off-by: Nikesh Oswal <Nikesh.Oswal@wolfsonmicro.com>