switch: arizona: Always defer micd handler to the workqueue
The micd_handler is potentially not thread-safe in the case where
its IRQ fires spuriously at the same time as one deferred to the
workqueue is also running. This can result in the first one to
start assuming that it has cancelled the timer but in fact the
timer was restarted during the race window.

To avoid this, always queue the micd handling on the workqueue.

Change-Id: I2775e92801d8e64dbfd14265e9538c58fdc49c0d
Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>