switch: arizona: force_micbias must switch micbias when changing detect mode
The previous behaviour of the driver when force_micbias was enabled was
to enable the MICBIAS pin for the initial micd_config mode and leave it
enabled through the entire mic detect process.

This won't work when there are multiple micd_configs (for different types
of probing) and they require different MICBIAS outputs.

This patch fixes this so that when the mic detect state machine moves to
the next micd_config mode it will disable any forced micbias for that mode
and then enable any forced micbias for the new mode.

Change-Id: Ic758ebb198c4b1f98f5753130356bfcf0a3c2722
Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>
Signed-off-by: bsapthagiri <sapthagiri.baratam@incubesol.com>