gpio: arizona: Activate arizona before handling gpios
GPIO registers are volatile so we must activate arizona
before attempting read/write on the register. Also when
we are modifying the register value we need to store a
copy of it in the cache so that the same can be
restored when the chip is reset or powered up.

Change-Id: Ie89de063d937684ff76053ad0b303f38ed134b7a
Signed-off-by: Nikesh Oswal <Nikesh.Oswal@wolfsonmicro.com>