ASoC: arizona: Don't call into wm_adsp driver from arizona driver
WM8997 and WM8998 don't have any DSPs so they don't require the
wm_adsp driver but do still use the arizona.c shared code.

Previously there was code in arizona.c that called wm_adsp_early_event()
and we were preventing a build error for WM8997/WM8998 by providing
a dummy implementation if the CONFIG_SND_SOC_WM_ADSP was not defined.
But this is not compatible with building as modules and leads to
a different build error because of the multiple definitions of
the function.

This patch removes the problem code from arizona.c. Since only
WM5102 and WM5110 are using that code, and it is only
a few lines, it can be moved into the codec drivers.

Change-Id: I4a6abb954f2549f408b1cf273b59949f35f33193
Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>