switch: madera: move pdata settings into separate struct
The pdata settings that only apply to the switch driver have been
moved out of the MFD pdata header into a separate struct in a
new header file. Also to allow for multiple accessory detects in
the future the main MFD pdata now contains an array of these new
structs.

Currently the switch driver still only implements support for a
single accessory detect.

Change-Id: I198f71b0b5d7124773919328ffe2518a5e6d672c
Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>