ASoC: madera: Make HP out status variables into arrays
Some codecs will support multiple "headphone" type outputs -
that is outputs that include accessory detection. This patch
makes the status variables for the hp output into arrays so
we can support multiple independant HP output states. The
hp_ena status does not need to be an array because it is a
set of bitflags for each output channel, matching register
0x400 Output_Enables_1.

Change-Id: I6aeff8dcf0ccce93748bc274966797d05b97cb72
Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>