switch: madera: move DT bindings into child nodes per output
To allow for codecs that have more than one accessory detect
this patch moves the DT bindings into child nodes so that a
node can be defined for each output on which accessory detect
must run.

Note that currently the driver only supports a single accessory
detect on OUT1 so any additional bindings are read but ignored.

Change-Id: I7e86cd6019c82880266c3888d84d6bf561850dec
Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>