ASoC: madera: Exit startup early if no runtime
Upstream commit 9b8ef9f6b3fc ("ASoC: dapm: Add startup & shutdown for dai_links")
added support for calling startup on CODEC to CODEC links, however this
is called with a NULL runtime pointer. There isn't really a sensible way
to pass a valid runtime pointer to a CODEC to CODEC link at the moment,
so we need to make the startup function safe for NULL runtimes.

This patch returns from the startup function early if there is no runtime,
this is perfectly safe as all the startup function does is set the PCM
constraints for user-space which aren't relevant to a CODEC to CODEC link.

Change-Id: I41d7c22662fd7c840efc8ff21579160d65633c6f
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>
Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>