ASoC: wm5110: Add support for audio trace firmware
Mainly this adds DAI links for the audio trace, however, it is also
necessary to update the data IRQ handler to check more cores. We have
the handler check every core so it should not be necessary to update
this function if more compressed firmwares are added in the future.

Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 31889507fd84a6a1c8dce36a214a070cec1fc559)

Change-Id: Ic51e0b47d97c595ea5f0c09cf2bcbcf614816601
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>