ASoC: madera: don't call into wm_adsp from madera.c
If madera.c calls into wm_adsp.c this has to be compiled out
if we ever have to build for a codec that doesn't have a DSP.
But this compiling-out causes problems when building as modules
because we can't rely on the WM_ADSP defines in a module build.
Since the affected code is fairly trivial, remove it from madera.c
and call from the codec drivers.

Change-Id: I4ea184828afc9c1640d6c231b36b340295b728f8
Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>