Revert "ASoC: arizona: cache the fll configuration"
This reverts commit e04abe46ad10b1f801b5ce435aa693058fe188e0.

The reverted patch was supposed to cache the last calculated FLL
settings and re-use them next time the FLL is enabled if the FLL
configuration hasn't changed. Unfortunately it was selecting the
previously cached settings even if the FLL configuration had been
changed.

The original patch wasn't necessary because the calculation time
is insignificant compared to the SPI writes, and regmap already
skips writing a register if its value hasn't changed. So reverting
the patch is better than fixing it since it reduces code complexity.

Change-Id: I4a8db322df322ee88ff9a12594987fc5ba6a90b0
Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>
Signed-off-by: Simon Trimmer <simont@opensource.wolfsonmicro.com>