ASoC: madera: Use mic widget for the HAPTICS widget
The HAPTICS widget will have enable_pin/disable_pin called on it, these
functions do not work well on siggen widgets because such widgets are
always on. This leads to odd behaviour where calling enable_pin on the
HAPTICS whilst it is part of a path will not cause the path to power
up. Fix this up by using a microphone widget for the siggen.

Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>
Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>

[ Probably not needed on 4.4 ]

Change-Id: I4ae3aba41215f09d691e8eb2a28b8c3727d1ff30
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>