ASoC: wm_adsp: Explicitly set the control access flags in all cases
In the case where the firmware does not tell us the access flags for the
control, we let ALSA select a default (READWRITE). But really we should
be applying the volatile flag in this case, as we will read the control
from the DSP if it is on in this case. This patch explicitly sets the
access flags in all cases.

Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit cc815c4b86d14630a7d4ca4f4c7d2c2acb88cb52)

Change-Id: I05da40c26eb944d880c4777bf77ded9944265bb0
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>