ASoC: madera: remove legacy Arizona FLL setup workarounds
The Madera driver has inherited from Arizona some complexity in the FLL
setup that only exists for legacy compatibility with early version of
the Arizona driver. This is irrelevant for Madera so this patch removes:

- using the syncclk settings as the refclk if we haven't been given any
  refclk settings

- defaulting the refclk source to be the 32k clock

The patch also updates the naming of the FLL defines used by the machine
driver to make it clearer whether the refclk or syncclk is being configured.

A small reordering of the defines is done to keep all the macros used by
the machine driver for configuring FLLs in a contiguous block.

Change-Id: I6b0304a6189f8dfd1506f803cd0d8b8b3cc3ed11
Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>