ASoC: madera: Expand inmode to include B mux settings
Worst case an input can have a maximum of four channel mode settings:
a left and right settings for each of the A and B sides of the input
mux. This patch expands inmode to have four values per input. For
pdata the inmode array is made two-dimensional to make the code
clearer. As device tree doesn't support multidimentional arrays the
DT is still a flat array but now has four cells per input.

The cs47l85 muxes are a little eccentric having only the left channel
muxed so the DT binding doc now uses the cs47l35 as the codec in the
example config to show a less confusing setup.

Change-Id: I9e016867e82b1b5b03c7b11368f66710a8e41cb9
Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>