ASoC: madera: mux channels must be ganged if INxA mode is digital
DMIC mode always requires signals from both the left and right channel
pins of an input, so in that case it would never be valid to have one
channel at 'A' and the other at 'B'. If the A input mode is DMIC the
mux control must always switch both channels together.

Change-Id: I47c310652641e72859d53734b62459b98664133c
Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>