switch: madera: Use hp impedance * 100 in external APIs
Use a precision of hundredths of an ohm when storing the HP impedance
or when passing it to the reading stage of the hpdet state machine.
This is achieved by passing ohms*100 as an integer, i.e. a 2 decimal
place fixed-point number.

Internally the driver is still using a precision of 1 ohm
but the APIs are now ready for higher precision if needed.

Change-Id: I57bfbebd0a4b23e44025dc65ee7b485c9ba3c774
Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>