ASoC: madera: remove use of fvco since it's an alias of fout
The FLL code used variables called fvco that actually held the
value of fout (which is fvco/3) - this is confusing.

Converting the fvco variables to actually hold fvco would result
in having to clutter the code with many divide-by-three, and is
redundant since the calculations really want to use fout.
So these variables have been renamed fout and the fvco member of
struct madera_fll is deleted.

Change-Id: I0ae1c0e9a66f4dd9eef992027a7454f8d7fb0713
Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>