mfd: madera: force reset if DCVDD did not power down during suspend
During suspend we mark regmap as cache-only, which means that resume
must sync the cache to ensure any register writes made during suspend
are written out to the codec. However, the regmap sync will skip any
cache entries that are at the register default value - this is only
safe if the codec actually was reset during suspend. If not, the
registers will not be in their default state and so any registers which
were changed in the cache to their default during suspend will not be
synced out to the codec.

This patch avoids this by checking in resume whether DCVDD really
powered off. If not it will force a hard or soft reset to ensure that
the registers are at defaults before syncing the cache.

Change-Id: I76161a4539097a7e8dd1683958181458dcb13d94
Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>