mfd: madera: Fix initialisation of the PM runtime
The PM runtime core by default assumes a chip is suspended when runtime
PM is enabled. Currently the madera driver gets around this by forcing
a madera_runtime_suspend() and then enabling runtime PM. This is a bit
ugly so instead this patch adds a call to pm_runtime_set_active to inform
the PM core that the device is starting up active. The PM enabling
happens around the IRQ init so that it is safe for PM operation to
enable/disable the IRQ.

Change-Id: Ib0a743dee23c28eed4c23623678dbd1b0d8b9c08
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>
Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>