ASoC: moon: Reduce headphone power up delay
The headphone output enable sequence is much faster on Moon only
requiring 6mS. This is applied only for Moon rather than for all chips
newer than Moon incase the delay increases on future chips. It is safer
to use the larger delay and reduce on a chip by chip basis.

Change-Id: Iead6ddf516bbb1796d08ba87792004408fe9c8cd
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>