ASoC: madera: Reduce cs47l90/91 headphone power up delay
The headphone output enable sequence is much faster on cs47l90 and
cs47l91 only requiring 6ms. This is applied only for those two codecs
rather than for all newer codecs in case the delay increases on future
codcs. It is safer to use the larger delay by default and reduce on a
chip-by-chip basis.

Change-Id: Ifd582bf432de8bd0da98545a75bd207688d15dea
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>
Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>