ASoC: madera: Create notifier to replace machine driver callbacks
This patch creates a notifier that can be used by future patches
for notifying the machine driver of events, rather than using a
collection of direct callbacks.

Any of the child drivers might want to send a notification so the
blocking_notifier_head is stored in struct madera, which is shared
by all the drivers. A helper function to call the notifiers is also
provided here.

The normal client of these notifications is the machine driver so
helper functions to register and unregister a callback are provided
that take a struct snd_soc_codec *.

Change-Id: Iab906addc97e78f09e57eb29e475f830615799b9
Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>