switch: treat hpdet < series resistance as a short
If there is an external series resistor on HPDET this must
be subtracted from any reading to get the actual impedance.

The code was trapping the case where the measurement was less
than the external resistor  but incorrectly just reporting the
measured value. It's entirely possible that due to component
tolerance and general variations a short or very low impedance
could be slightly below the declared external resistance, but
in those cases it would incorrectly be reported as a much larger
impedance.

If the measured value is less than the declared series resistance
it should be treated as a short circuit.

Change-Id: Ifeb44a211845cf8bf9a6a7ad599604e15b8bb2af
Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>