switch: madera: remove MADERA_HP_SHORT_IMPEDANCE
Remove the define MADERA_HP_SHORT_IMPEDANCE from the pdata
header. It's meaningless to define it here because it doesn't
necessarily represent the actual threshold we're using for
short circuit (that can be overridden).

Its only purpose was to limit the lower end of the custom short
circuit threshold to not be less than the hardware can measure
and this has been replaced by MADERA_HP_SHORT_IMPEDANCE_MIN
in switch-madera.c

Change-Id: I21924129fdc829064e6a7c0bd1866f05c5bd75d6
Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>