ASoC: madera: Fix handling of snd_soc_update_bits return value
The snd_soc_update_bits() function returns 0 if no bits changed
and 1 if bits were changed successfully. For cases where this
return value was being used as the return of the calling function
this would fail if the next caller up was assuming non-zero means
error. So this patch converts all "success" returns from
snd_soc_update_bits() to return 0.

Change-Id: Ibbcf580cfa849fdc58e1ed16b06298c8b934dc17
Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>