switch: madera: Defer probe if irqchip driver hasn't probed
We depend on the madera irqchip driver to have probed to provide
us with interrupts, so defer the switch driver probe if the irqchip
driver isn't ready yet. The madera->irq_dev member is set by the
irqchip driver only when it has successfully completed probe so
this can be used to perform an early EPROBE_DEFER exit.

Change-Id: Id4cfcfa886fe6f0aed3dde11e317841f4ca1adca
Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>