mfd: madera: Make gpio_is_valid() work on unset pdata
We use a value of 0 in pdata to indicate that the reset GPIO has not
been assigned (since C sets undefined members of a struct to 0). But
0 is a valid GPIO so gpio_is_valid() would give the wrong answer.
Since it's potentially confusing the gpio_is_valid() won't work, this
patch converts a 0 pdata value for reset into -1 to be compatible with
normal GPIO numbering convention.

Change-Id: I091bd88a7b070da881ac6053712b873a9efb1721
Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>