ASoC: wm_adsp: Only attempt to populate regions if there are compress caps
If a firmware with a host buffer was named as a selection that does
not define any compressed caps or host buffer regions the driver
would attempt to initialize the host buffer and then hit a NULL
pointer dereference when it tries to access the caps and host regions.

This patch takes additional upstream check that there are valid caps
before populating the buffer info.

Change-Id: Ia22192e77f59a8870da8016c0edfe40a13280666
Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>