regulator: madera-ldo1: Do not mark as always-on if no LDOENA
This partially reverts commit 7b4a68fcef6f353dd0dd698f097d9db286c7a55c.
"regulator: madera-ldo1: Mark as always-on if no LDOENA"

The intention was to indicate that the regulator cannot be turned off
if it does not have an LDOENA, but the always_on flag actually means
"must not be turned off", not "unable to turn off" so it's wrong to
mark it always on unless we want to keep the supply permanently
enabled - even if we don't haev an LDOENA the supply could still be
turned off by disabling the parent regulator.

The original patch also set the ldo1->init_data to be the actual
init data in use. This was done to simplify the implementation of
the always_on handling, but it's not actually part of this so it
has not been reverted. It's reasonable that as we're already using
space for the ldo1->init_data it should be the actual init_data.

Change-Id: I4d9a4464aae9a13d1234e1db58d3533aad0acf91
Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>