ASoC: wm_adsp: Move control cache sync out of preloader
As part of the work to download firmware before the audio path is brought
up the DSP will be put into a low power state between downloading firmware
to the core and starting it running. This will mean that the firmware ALSA
controls are not accessible in the hardware during this period of
time.

To prepare for this change we gate access to the hardware in the ALSA
control handlers on the DSP being running rather than simply booted and
move the synchronisation of the control caches out of the preloader delayed
work and into the main DAPM thread after the DSP will have been brought out
of its low power state.

Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit cef45771c141fdccebe4cb7e0ce79f4687275494)

Change-Id: I540189049706fc226495972ed3e731e9414bed23
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>